# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/models":{"post":{"operationId":"ModelController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateModelDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"example":{"data":{"id":1,"name":"Model"}}}}}}},"summary":"Создать модель","tags":["Model"]},"get":{"operationId":"ModelController_getAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"data":[]}}}}}},"summary":"Все модели","tags":["Model"]}},"/api/models/all/include/source":{"get":{"operationId":"ModelController_getAllIncludeSouce","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"data":[]}}}}}},"summary":"Все модели с источниками","tags":["Model"]}},"/api/models/count":{"get":{"operationId":"ModelController_getCount","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"data":{"count":42}}}}}}},"summary":"Количество моделей","tags":["Model"]}},"/api/models/notation/{id}":{"get":{"operationId":"ModelController_notation","parameters":[{"name":"id","required":true,"in":"path","schema":{"example":"1","type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"data":[]}}}}},"404":{"description":"Нотация не найдена"}},"summary":"Нотация по ID источника","tags":["Model"]}},"/api/models/{id}":{"get":{"operationId":"ModelController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"example":"1","type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"data":{"id":1,"name":"Model"}}}}}},"404":{"description":"Модель не найдена"}},"summary":"Модель по ID","tags":["Model"]}},"/api/sources":{"post":{"operationId":"SourcesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSourceDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"example":{"data":{"id":1,"name":"string"}}}}}}},"summary":"Создать источник","tags":["Sources"]},"get":{"operationId":"SourcesController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"data":[{"id":1,"name":"string"},{"id":2,"name":"string"}]}}}}}},"summary":"Все источники","tags":["Sources"]}},"/api/sources/full":{"get":{"operationId":"SourcesController_full","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"data":[{"id":1,"name":"string","key":{"id":1,"sourceId":1,"apiKey":"string","isActive":true,"createdAt":"string","updatedAt":"string"},"models":[]},{"id":1,"name":"Gemini","key":null,"models":[]}]}}}}}},"summary":"Все источники с моделями","tags":["Sources"]}},"/api/sources/full/{id}":{"get":{"operationId":"SourcesController_fullById","parameters":[{"name":"id","required":true,"in":"path","schema":{"example":"1","type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"data":{"id":1,"name":"OpenAI","models":[{"id":1,"name":"gpt-4"}]}}}}}},"404":{"description":"Источник не найден"}},"summary":"Источник с моделями по ID","tags":["Sources"]}},"/api/sources/{id}":{"get":{"operationId":"SourcesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"example":"1","type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"data":{"id":1,"name":"OpenAI"}}}}}},"404":{"description":"Источник не найден"}},"summary":"Источник по ID","tags":["Sources"]},"delete":{"operationId":"SourcesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"example":"1","type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"data":{"id":1,"name":"string"}}}}}},"404":{"description":"Источник не найден"}},"summary":"Удалить источник","tags":["Sources"]}},"/api/keys":{"post":{"description":"Создает новый API ключ","operationId":"KeysController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateKeyDto"}}}},"responses":{"201":{"description":"Ключ успешно создан","content":{"application/json":{"schema":{"example":{"data":{"id":"uuid-123","name":"My API Key","key":"sk-1234567890abcdef","createdAt":"2025-12-10T12:00:00Z"}}}}}},"400":{"description":"Некорректные данные"}},"summary":"Создать ключ","tags":["Keys"]},"get":{"operationId":"KeysController_findAll","parameters":[],"responses":{"200":{"description":"Список ключей","content":{"application/json":{"schema":{"example":{"data":[{"id":"uuid-1","name":"Key 1","createdAt":"2025-12-10T12:00:00Z"},{"id":"uuid-2","name":"Key 2","createdAt":"2025-12-10T12:01:00Z"}]}}}}}},"summary":"Получить все ключи","tags":["Keys"]}},"/api/keys/{id}":{"get":{"operationId":"KeysController_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"ID ключа","schema":{"example":"uuid-123","type":"string"}}],"responses":{"200":{"description":"Ключ найден","content":{"application/json":{"schema":{"example":{"data":{"id":"uuid-123","name":"My API Key","key":"sk-1234567890abcdef","createdAt":"2025-12-10T12:00:00Z"}}}}}},"404":{"description":"Ключ не найден"}},"summary":"Получить ключ по ID","tags":["Keys"]},"patch":{"operationId":"KeysController_update","parameters":[{"name":"id","required":true,"in":"path","description":"ID ключа","schema":{"example":"uuid-123","type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateKeyDto"}}}},"responses":{"200":{"description":"Ключ обновлен","content":{"application/json":{"schema":{"example":{"data":{"id":"uuid-123","name":"Updated Key Name","updatedAt":"2025-12-10T12:05:00Z"}}}}}},"404":{"description":"Ключ не найден"}},"summary":"Обновить ключ","tags":["Keys"]},"delete":{"operationId":"KeysController_remove","parameters":[{"name":"id","required":true,"in":"path","description":"ID ключа","schema":{"example":"uuid-123","type":"string"}}],"responses":{"200":{"description":"Ключ удален","content":{"application/json":{"schema":{"example":{"data":{"message":"Ключ успешно удален"}}}}}},"404":{"description":"Ключ не найден"}},"summary":"Удалить ключ","tags":["Keys"]}},"/api/ai/perplexity/{model}/static":{"post":{"description":"Отправить сообзение для провайдера Perplexity","operationId":"AiController_send","parameters":[{"name":"model","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendDto"}}}},"responses":{"2XX":{"description":"","content":{"application/json":{"schema":{"example":{"data":{"content":"content","role":"role"}}}}}}},"summary":"Отправить сообщение","tags":["AI"]}},"/api/ai/perplexity/{model}/stream":{"post":{"operationId":"AiController_ppls","parameters":[{"name":"model","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["AI"]}},"/api/messages":{"post":{"operationId":"MessagesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMessageDto"}}}},"responses":{"201":{"description":"Сообщение успешно создано"},"400":{"description":"Некорректные данные"}},"summary":"Создать новое сообщение","tags":["Messages"]}},"/api/messages/initial":{"post":{"operationId":"MessagesController_initial","parameters":[],"requestBody":{"required":true,"description":"chatId обязателен, content — текст сообщения","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMessageDto"}}}},"responses":{"201":{"description":"Начальное сообщение создано"},"400":{"description":"Некорректные данные"}},"summary":"Создать начальное сообщение для чата","tags":["Messages"]}},"/api/messages/{id}":{"get":{"operationId":"MessagesController_allForChat","parameters":[{"name":"id","required":true,"in":"path","description":"ID чата","schema":{"type":"number"}}],"responses":{"200":{"description":"Список сообщений для указанного чата"},"404":{"description":"Чат не найден"}},"summary":"Получить все сообщения для чата","tags":["Messages"]}},"/api/chats":{"post":{"operationId":"ChatsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateChatDto"}}}},"responses":{"201":{"description":"Чат создан"},"400":{"description":"Неверные данные"}},"summary":"Создать новый чат","tags":["chats"]}},"/api/chats/all":{"get":{"operationId":"ChatsController_findMany","parameters":[],"responses":{"200":{"description":"Список чатов"}},"summary":"Получить список всех чатов","tags":["chats"]}},"/api/chats/{id}":{"get":{"operationId":"ChatsController_chat","parameters":[{"name":"id","required":true,"in":"path","description":"ID чата","schema":{"type":"number"}}],"responses":{"200":{"description":"Чат найден"},"404":{"description":"Чат не найден"}},"summary":"Получить чат по ID","tags":["chats"]},"delete":{"operationId":"ChatsController_deleteChat","parameters":[{"name":"id","required":true,"in":"path","description":"ID чата","schema":{"type":"number"}}],"responses":{"200":{"description":"Чат удалён"},"404":{"description":"Чат не найден"}},"summary":"Удалить чат по ID","tags":["chats"]}},"/api/chats/search":{"get":{"operationId":"ChatsController_findByTitle","parameters":[{"name":"title","required":false,"in":"query","description":"Подстрока для поиска по названию чата","schema":{"type":"string"}}],"responses":{"200":{"description":"Список подходящих чатов"}},"summary":"Найти чаты по title","tags":["chats"]}},"/api/chats/select/{id}":{"get":{"operationId":"ChatsController_select","parameters":[{"name":"id","required":true,"in":"path","description":"ID чата","schema":{"type":"number"}}],"responses":{"200":{"description":"Чат выбран"},"404":{"description":"Чат не найден"}},"summary":"Выбрать чат (например, сделать активным)","tags":["chats"]}}},"info":{"title":"Nova api","description":"Api description","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateModelDto":{"type":"object","properties":{"name":{"type":"string","description":"Название модели","example":"gemini-2.5-flash","maxLength":20},"sourceId":{"type":"number","description":"ID источника модели","example":1,"minimum":1}},"required":["name","sourceId"]},"CreateSourceDto":{"type":"object","properties":{"name":{"type":"string","description":"Имя источника","example":"Gemini"}},"required":["name"]},"CreateKeyDto":{"type":"object","properties":{"sourceId":{"type":"number","description":"Для какого источника создавать ApiKey","minimum":1,"example":1},"apiKey":{"type":"string","description":"Api ключ","example":"pplx-**********"},"isActive":{"type":"boolean","description":"Являеться на данный момент ключ активным.","example":true}},"required":["sourceId","apiKey"]},"UpdateKeyDto":{"type":"object","properties":{"sourceId":{"type":"number","description":"Для какого источника создавать ApiKey","minimum":1,"example":1},"apiKey":{"type":"string","description":"Api ключ","example":"pplx-**********"},"isActive":{"type":"boolean","description":"Являеться на данный момент ключ активным.","example":true}}},"SendDto":{"type":"object","properties":{"messages":{"type":"array","description":"Масив сообщений","maxLength":1000,"example":"Что такое Linux"}},"required":["messages"]},"CreateMessageDto":{"type":"object","properties":{"chatId":{"type":"number","description":"Айди чата для которого сообщение","example":1,"minimum":1},"content":{"type":"string","description":"Текст сообщения","example":"Что такое Nixos?"},"role":{"type":"string","description":"Роль для текущего сообщения","example":"user","enum":["user","assistant","system"]}},"required":["chatId","content"]},"CreateChatDto":{"type":"object","properties":{"provider":{"type":"string","description":"Провайдер в формате `name/model`. Например `gemini/gemini-2.5-flash`","example":"perplexity/sonar"},"title":{"type":"string","description":"Начальное сообщение для чата.","example":"Что такое Nixos?"}},"required":["provider","title"]}}}}